requirements.txt
Django>=3.1.6,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.7.5,<2.8.0
django-crispy-forms>=1.11.1,<1.11.9
flake8>=3.8.4,<3.9.0
django-filter~=2.4.0